Release/v0.4

See merge request compare-it/compare-it-back!12

+++ src/main/java/com/pepit/service/impl/CompanyServiceImpl.java
@@ -5,11 +5,15 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.mysql.cj.xdevapi.*;
 import com.pepit.constants.TypeModelPropertyEnum;
+import com.pepit.converters.CompanyConverter;
+import com.pepit.dto.CompanyDto;
 import com.pepit.exception.DataProvidedException;
 import com.pepit.exception.InputException;
 import com.pepit.exception.ReferentielRequestException;
+import com.pepit.model.Company;
 import com.pepit.model.Model;
 import com.pepit.model.WebsiteConfiguration;
+import com.pepit.repository.CompanyRepository;
 import com.pepit.repository.ProductRepositoryCustom;
 import com.pepit.service.CompanyService;
 import com.pepit.service.WebsiteConfigurationService;
@@ -26,10 +30,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.StringReader;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Service
 @Slf4j
@@ -37,11 +38,16 @@ public class CompanyServiceImpl implements CompanyService {
 
     private ProductRepositoryCustom productRepository;
     private WebsiteConfigurationService websiteConfigurationService;
+    private CompanyRepository companyRepository;
+    private CompanyConverter companyConverter;
+
 
     @Autowired
-    public CompanyServiceImpl(ProductRepositoryCustom productRepository, WebsiteConfigurationService websiteConfigurationService) {
+    public CompanyServiceImpl(ProductRepositoryCustom productRepository, WebsiteConfigurationService websiteConfigurationService, CompanyRepository companyRepository, CompanyConverter companyConverter) {
         this.productRepository = productRepository;
         this.websiteConfigurationService = websiteConfigurationService;
+        this.companyRepository = companyRepository;
+        this.companyConverter = companyConverter;
     }
 
     public String fromUrlToDb(String url, String supplierId, String type) {
@@ -130,15 +136,18 @@ public class CompanyServiceImpl implements CompanyService {
                 dbDocList.add(outerObject);
             }
         }
-        //on le convertit en tableau car dependance du add mysql
-        DbDoc[] docs = dbDocList.toArray(new DbDoc[dbDocList.size()]);
 
-        //TODO Utiliser le generateur de QUERY
-        productRepository.removeDoc("supplierId = " + supplierId + " and type = '" + typeProduit.replace("\"", "") + "'");
-        productRepository.addDoc(docs);
+        if (!dbDocList.isEmpty()) {
+            //on le convertit en tableau car dependance du add mysql
+            DbDoc[] docs = dbDocList.toArray(new DbDoc[dbDocList.size()]);
 
-        //On évalue les bornes et valeurs des proprietes du modele selon les données en base
-        recalculMinMaxValue(typeProduit);
+            //TODO Utiliser le generateur de QUERY
+            productRepository.removeDoc("supplierId = " + supplierId + " and type = '" + typeProduit.replace("\"", "") + "'");
+            Iterator<Warning> result = productRepository.addDoc(docs);
+
+            //On évalue les bornes et valeurs des proprietes du modele selon les données en base
+            recalculMinMaxValue(typeProduit);
+        } else throw new DataProvidedException("FAILED : Aucun produit ajouté");
 
         if (!refusedRecords.isEmpty()) {
             log.info("Lignes rejettées: " + refusedRecords.toString());
@@ -179,10 +188,12 @@ public class CompanyServiceImpl implements CompanyService {
         try {
             WebsiteConfiguration wsc = websiteConfigurationService.findOneById(1);
             wsc.getModelByTechnicalName(typeProduit).getModelProperties().forEach(prop -> {
-                if (prop.getType().equals(TypeModelPropertyEnum.NUMERIC))
-                    productRepository.updateBornes(wsc.getModelByTechnicalName(typeProduit), prop.getTechnicalName());
-                else
-                    prop.setValues(productRepository.listeDistinct(typeProduit, prop.getTechnicalName()));
+                if (prop.filtrable || prop.filtrableAdvanced) {
+                    if (prop.getType().equals(TypeModelPropertyEnum.NUMERIC))
+                        productRepository.updateBornes(wsc.getModelByTechnicalName(typeProduit), prop.getTechnicalName());
+                    else
+                        prop.setValues(productRepository.listeDistinct(typeProduit, prop.getTechnicalName()));
+                }
             });
             websiteConfigurationService.save(wsc);
         } catch (ReferentielRequestException e) {
@@ -230,5 +241,11 @@ public class CompanyServiceImpl implements CompanyService {
                 .add("type", new JsonString().setValue(typeProduit))
                 .add("properties", properties);
         return outerObject;
+    }
+
+    @Override
+    public CompanyDto create(CompanyDto companyDto) {
+        Company company = companyConverter.dtoToEntity(companyDto);
+        return companyConverter.entityToDto(this.companyRepository.save(company));
     }
 }
